test(frontend): the company page's backstops are stated once, and a walk's failure is waited for
The task-tab suite moved out of a 2,668-line file that the frontend rules cap at
1,000 and explicitly say not to grow. The whole describe went, not the one test
this branch added: moving a single test would have cost the same harness for a
quarter of the benefit.

Splitting it duplicated the harness, which is the more interesting half. What
`CompanyScreen` reads was then stated twice — and the old file already stated it
twice more, in an enrichment stub and a deep-read stub that had each restated the
same five page-shell answers by hand. One module answers the wire now and all
four sites read it. The proof the extraction found the real shape: afterwards the
roll-up, brief and strength defaults had no reference left in either test file,
because they existed only to feed the copies. Both suites shrank by 292 lines
against 196 shared, and the test count is unchanged at 75, so nothing was lost or
quietly skipped.

`render` stays local deliberately. It is identical in 56 files and has no drift
coupling to what this page reads, so a company-named module is the wrong home for
it; one authority for it belongs beside the app harness, as its own change.

**A test helper was also driving a second input device.** `openTasksTab` used the
static `userEvent`, so four tests each built an implicit instance per
interaction — and a second instance forgets which keys and buttons the first left
held. All four pass the one they set up.

**And one assertion could not have failed.** After a later page fails, the
Documents tab renders markup identical to before the click — same rows, same
partial-list sentence, same enabled button — because failure is reserved for
having nothing. So no wait over the output could distinguish "the second page
failed and the first survived" from "nothing has happened yet". The test now
holds the second response itself and waits on the states either side of it: the
button disabled while the walk is out, then not, and only then that the rows
survived. Proven both ways — against an implementation that drops the rows on a
failed page the new form fails and the old form passes.
